update ctf filter

+++ libEM/sparx/fourierfilter.cpp
@@ -396,11 +396,19 @@ EMData* Processor::EMFourierFilterFunc(EMData * fimage, Dict params, bool doInPl
 							break;
 						case CTF_:
 						  if(ny>1 && nz<=1 )						  					
-						   { ak=sqrt((float)(jx/lsd2*jx/lsd2+jy/nyp2*jy/nyp2))/ps/2.0f;}
+
+						   { ak=sqrt(static_cast<float>(jx)/lsd2*static_cast<float>(jx)/lsd2 +
+						        static_cast<float>(jy)/nyp2*static_cast<float>(jy)/nyp2)/ps/2.0f;}
+
 						  else	if(ny<=1)
-						  	     { ak=sqrt((float)(jx/lsd2*jx/lsd2))/ps/2.0f;}
+
+						  	     { ak=sqrt(static_cast<float>(jx)/lsd2*static_cast<float>(jx)/lsd2)/ps/2.0f;}						  	    
+
 						  else  if(nz>1)
-						     { ak=sqrt((float)(jx/lsd2*jx/lsd2+jy/nyp2*jy/nyp2+jz/nzp2*jz/nzp2))/ps/2.0f;}          
+						     {
+                                                       ak=sqrt(static_cast<float>(jx)/lsd2*static_cast<float>(jx)/lsd2 +
+						               static_cast<float>(jy)/nyp2*static_cast<float>(jy)/nyp2 +
+							       static_cast<float>(jz)/nzp2*static_cast<float>(jz)/nzp2)/ps/2.0f;}          						   	
 						       				
 						  fp->cmplx(ix,iy,iz) *= Util::tf(dz,ak,12.398f/sqrt(lambda *(1022.f+lambda)),cs*1.0e-7f,atan(wgh/(1.0-wgh)),b_factor,sign);
